+++ Sources/StreamChatSwiftUI/Utils/StringExtensions.swift
@@ -112,3 +112,17 @@ extension String {
     
     static let unknownMessageId = "unknown"
 }
+
+extension String {
+    func ranges<S: StringProtocol>(of string: S, options: String.CompareOptions = []) -> [Range<String.Index>] {
+        var result: [Range<Index>] = []
+        var startIndex = self.startIndex
+        while startIndex < endIndex, let range = self[startIndex...].range(of: string, options: options) {
+            result.append(range)
+            startIndex = range.lowerBound < range.upperBound
+                ? range.upperBound
+                : index(range.lowerBound, offsetBy: 1, limitedBy: endIndex) ?? endIndex
+        }
+        return result
+    }
+}