+++ StreamChatSwiftUITests/Tests/Utils/StringExtensions_Tests.swift
@@ -62,4 +62,24 @@ class String_Extensions_Tests: XCTestCase {
         XCTAssert("example".isURL == false)
         XCTAssert("invalid_url".isURL == false)
     }
+    
+    func testRangesOfString() {
+        let mention = "@Martin"
+        let string = "Hey \(mention), how are you?"
+        let result = string
+            .ranges(of: mention, options: [.caseInsensitive])
+            .map { NSRange($0, in: string) }
+            .first
+        XCTAssertEqual(result, NSRange(location: 4, length: 7))
+    }
+    
+    func testRangesOfStringNotFound() {
+        let string = "Hey @Martin, how are you?"
+        let mention = "@Alexey"
+        let result = string
+            .ranges(of: mention, options: [.caseInsensitive])
+            .map { NSRange($0, in: string) }
+            .first
+        XCTAssertEqual(result, nil)
+    }
 }